add entry point; reorganize scripts
