+++ bin/run.js
@@ -1,0 +1,3 @@
+const F = require("../lib/ftp-deploy.js");
+
+F.ftpDeploy(process.argv.slice(2))
