+++ lib/ftp-deploy.js
@@ -3,12 +3,6 @@
  * Existing site will be moved to a backup directory.
  */
 
-// Requires Dev dependencies:
-//  npm install file-system
-//  npm install promise-ftp
-//  npm install bluebird
-//  npm install minimalist
-
 var Promise = require("bluebird");
 var fileSystem = require("file-system");
 var PromiseFtp = require("promise-ftp");
@@ -113,7 +107,7 @@ module.exports = {
       .then(backupPath => {
         // Backup production html directory.
         console.log("Renaming", opts.domain + "html/", "to", backupPath);
-        return this.ftp.rename(opts.domain + "html/", backupPath);
+        return ftp.rename(opts.domain + "html/", backupPath);
       })
       .then(message => {
         // Deploy stage directory.
@@ -124,7 +118,7 @@ module.exports = {
           "to",
           opts.domain + "html/"
         );
-        return this.ftp.rename(opts.domain + opts.stageDir, opts.domain + "html/");
+        return ftp.rename(opts.domain + opts.stageDir, opts.domain + "html/");
       })
       .then(message => {
         // Log final message.
@@ -184,5 +178,3 @@ module.exports = {
     return this.deploy(opts);
   }
 };
-
-module.ftpDeploy(process.argv.slice(2))
