+++ package.json
@@ -1,47 +1,50 @@
 {
-    "name": "ftp-deploy",
-    "version": "0.1.0",
-    "description": "Deploy a static site via ftp.",
-    "main": "index.js",
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/gtwiggs/ftp-deploy"
-    },
-    "keywords": [
-        "ftp",
-        "deploy",
-        "site"
-    ],
-    "author": {
-        "name": "Glenn Twiggs",
-        "email": "info@glenntwiggs.com",
-        "url": "http://glenntwiggs.com"
-    },
-    "readmeFilename": "README.md",
-    "dependencies": {
-        "bluebird": "^3.5.3",
-        "file-system": "^2.2.2",
-        "minimalist": "^1.0.0",
-        "promise-ftp": "^1.3.5",
-        "yarn": "^1.12.3"
-    },
-    "devDependencies": {
-        "covert": "^1.1.0",
-        "sinon": "^7.2.2",
-        "tap": "^12.1.1",
-        "tape": "^4.9.1"
-    },
-    "optionalDependencies": {},
-    "engines": {
-        "node": "*"
-    },
-    "bugs": {
-        "url": "https://github.com/gtwiggs/ftp-deploy/issues"
-    },
-    "homepage": "https://github.com/gtwiggs/ftp-deploy",
-    "scripts": {
-        "test": "tap test/*.js",
-        "coverage": "covert test/*.js"
-    },
-    "license": "MIT"
+  "name": "ftp-deploy",
+  "version": "0.1.0",
+  "description": "Deploy a static site via ftp.",
+  "main": "lib/ftp-deploy.js",
+  "bin": {
+    "dewploy": "bin/run.js"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/gtwiggs/ftp-deploy"
+  },
+  "keywords": [
+    "ftp",
+    "deploy",
+    "site"
+  ],
+  "author": {
+    "name": "Glenn Twiggs",
+    "email": "info@glenntwiggs.com",
+    "url": "http://glenntwiggs.com"
+  },
+  "readmeFilename": "README.md",
+  "dependencies": {
+    "bluebird": "^3.5.3",
+    "file-system": "^2.2.2",
+    "minimalist": "^1.0.0",
+    "promise-ftp": "^1.3.5",
+    "yarn": "^1.12.3"
+  },
+  "devDependencies": {
+    "covert": "^1.1.0",
+    "sinon": "^7.2.2",
+    "tap": "^12.1.1",
+    "tape": "^4.9.1"
+  },
+  "optionalDependencies": {},
+  "engines": {
+    "node": "*"
+  },
+  "bugs": {
+    "url": "https://github.com/gtwiggs/ftp-deploy/issues"
+  },
+  "homepage": "https://github.com/gtwiggs/ftp-deploy",
+  "scripts": {
+    "test": "tap test/*.js",
+    "coverage": "covert test/*.js"
+  },
+  "license": "MIT"
 }
